Migrate User and Asset List to livewire in asset creation form. Change styles in most views related to button placements

+++ app/Http/Livewire/ThreatControlsManage.php
@@ -7,10 +7,12 @@ use App\Models\Threat;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Contracts\View\View;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Livewire\Component;
 
 class ThreatControlsManage extends Component
 {
+    use AuthorizesRequests;
     public $threat;
     public $control;
 
@@ -21,6 +23,7 @@ class ThreatControlsManage extends Component
 
     public function render(): Factory|View|Application
     {
+        $this->authorize("update", $this->threat);
         $controls = $this->threat->controls();
         $all_controls = Control::whereNotIn("id", $controls->pluck("control_id")->toArray())->get();
         return view('livewire.threat-controls-manage', [
@@ -32,6 +35,7 @@ class ThreatControlsManage extends Component
 
     public function addControl()
     {
+        $this->authorize("update", $this->threat);
         if (!empty($this->control)) {
             $this->threat->controls()->attach($this->control);
             $this->control = null;
@@ -40,6 +44,7 @@ class ThreatControlsManage extends Component
 
     public function removeControl($id)
     {
+        $this->authorize("update", $this->threat);
         $this->threat->controls()->detach($id);
     }
 }
